+++ wazo_confd/http_server.py
@@ -1,4 +1,4 @@
-# Copyright 2016-2019 The Wazo Authors  (see the AUTHORS file)
+# Copyright 2016-2020 The Wazo Authors  (see the AUTHORS file)
 # SPDX-License-Identifier: GPL-3.0-or-later
 
 import os
@@ -117,6 +117,9 @@ class HTTPServer:
         bind_addr = (self.config['listen'], self.config['port'])
         self.server = wsgi.WSGIServer(bind_addr=bind_addr, wsgi_app=wsgi_app)
         if self.config['certificate'] and self.config['private_key']:
+            logger.warning(
+                'Using service SSL configuration is deprecated. Please use NGINX instead.'
+            )
             self.server.ssl_adapter = http_helpers.ssl_adapter(
                 self.config['certificate'], self.config['private_key']
             )